Refactor Tests + Demos: split + decompose
Tests:
- ElicitationTests (279 body) → +ElicitationSchemaTests.
- CalculatorMockTests (313) → +CalculatorMockErrorTests.
- URITemplateValidatorTests (356, file 448) → +Level + +Character
  suites.
- ComplexTypesTests (707) → +Array + +Object suites.
- MCPToolTests (695, func@196 cc 14) → +TestFixtures + +SchemaTests;
  testBasicFunctionality decomposed into 4 private assert helpers.
- HTTPTransportTests (908, body 726) → +TestHelpers + +Legacy +
  +Resume + +OpenAPI suites. ResumableServer + Box widened to internal
  (renamed HTTPTransportBox / HTTPTransportStreamCapture).
- MCPServerTests, ProxyGeneratorResourcePromptTests,
  ProxyGeneratorDefaultValueTests, OpenAPISpecTests: long test bodies
  decomposed into focused private helpers (makeInitializeRequest /
  extractInitializeResponseResult / assertInitializeCapabilities,
  makeDemoResources / makeDemoTemplates / makeDemoPrompts,
  makeDefaultsSchema, etc).

Demos:
- DemoServer.swift (747, actor 520) → tool bodies shrunk via
  Self.logCall; schema builders + describe…Response + message(for:)
  moved to DemoServer+ElicitationSchemas.swift.
- ProxyDemoCLI.swift (427, struct 285, func@267 cc 22) → all
  formatValue / formatArray / formatObject / escapeString moved to
  ProxyDemoFormatter enum; formatPrimitive split into formatScalar +
  formatFoundationLiteral.
- Demos/SwiftMCPDemo/Commands/HTTPSSECommand (func@124 cc 13) and
  SwiftMCPIntentsDemo/Commands/HTTPSSECommand (func@70 cc 14) →
  decomposed into configureAuthentication, configureOAuth,
  logOAuthConfiguration, makeTokenAuthorizationHandler,
  startTCPTransportIfNeeded, setupSignalHandling.

No `swiftlint:disable` for structural rules. Same 415 tests run.
Build clean.
