+++ Demos/SwiftMCPIntentsDemo/Commands/HTTPSSECommand.swift
@@ -83,68 +83,11 @@ final class HTTPSSECommand: AsyncParsableCommand {
         )
 
         let transport = HTTPSSETransport(server: server, port: port)
-
-        if let oauthConfigPath = oauth {
-            do {
-                let jsonConfig = try JSONOAuthConfiguration.load(from: oauthConfigPath)
-                let oauthConfig = try jsonConfig.toOAuthConfiguration()
-                transport.oauthConfiguration = oauthConfig
-
-                print("OAuth validation enabled with issuer: \(jsonConfig.issuer)")
-                if jsonConfig.transparentProxy == true {
-                    print("  Transparent proxy mode: enabled (server acts as OAuth provider)")
-                }
-                if jsonConfig.introspectionEndpoint == nil {
-                    print("  JWT validation: enabled (no introspection endpoint provided)")
-                } else {
-                    print("  Token introspection: enabled")
-                }
-                if let audience = jsonConfig.audience {
-                    print("  Expected audience: \(audience)")
-                }
-                if let clientID = jsonConfig.clientID {
-                    print("  Expected client ID: \(clientID)")
-                }
-            } catch {
-                print("Error loading OAuth configuration: \(error)")
-                throw ExitCode.failure
-            }
-        } else if let requiredToken = token {
-            transport.authorizationHandler = { token in
-                guard let token else {
-                    return .unauthorized("Missing bearer token")
-                }
-
-                guard token == requiredToken else {
-                    return .unauthorized("Invalid bearer token")
-                }
-
-                return .authorized
-            }
-            print("Simple token validation enabled")
-        } else {
-            transport.authorizationHandler = { _ in
-                return .authorized
-            }
-            print("No authentication configured - all requests will be accepted")
-        }
-
+        try configureAuthentication(on: transport)
         transport.serveOpenAPI = openapi
 
-        var tcpTransport: TCPBonjourTransport?
-        if tcp {
-            let transport = TCPBonjourTransport(server: server)
-            try await transport.start()
-            tcpTransport = transport
-            print("MCP Server \(server.serverName) started with TCP+Bonjour transport")
-        }
-
-        if let tcpTransport {
-            signalHandler = SignalHandler(transports: [transport, tcpTransport])
-        } else {
-            signalHandler = SignalHandler(transport: transport)
-        }
-        await signalHandler?.setup()
+        let tcpTransport = try await startTCPTransportIfNeeded(server: server)
+        await setupSignalHandling(httpTransport: transport, tcpTransport: tcpTransport)
 
         do {
             try await transport.run()
@@ -154,5 +97,80 @@ final class HTTPSSECommand: AsyncParsableCommand {
             }
             throw error
         }
+    }
+
+    private func configureAuthentication(on transport: HTTPSSETransport) throws {
+        if let oauthConfigPath = oauth {
+            try configureOAuth(on: transport, configPath: oauthConfigPath)
+        } else if let requiredToken = token {
+            transport.authorizationHandler = Self.makeTokenAuthorizationHandler(requiredToken: requiredToken)
+            print("Simple token validation enabled")
+        } else {
+            transport.authorizationHandler = { _ in .authorized }
+            print("No authentication configured - all requests will be accepted")
+        }
+    }
+
+    private func configureOAuth(on transport: HTTPSSETransport, configPath: String) throws {
+        do {
+            let jsonConfig = try JSONOAuthConfiguration.load(from: configPath)
+            transport.oauthConfiguration = try jsonConfig.toOAuthConfiguration()
+            Self.logOAuthConfiguration(jsonConfig)
+        } catch {
+            print("Error loading OAuth configuration: \(error)")
+            throw ExitCode.failure
+        }
+    }
+
+    private static func logOAuthConfiguration(_ jsonConfig: JSONOAuthConfiguration) {
+        print("OAuth validation enabled with issuer: \(jsonConfig.issuer)")
+        if jsonConfig.transparentProxy == true {
+            print("  Transparent proxy mode: enabled (server acts as OAuth provider)")
+        }
+        if jsonConfig.introspectionEndpoint == nil {
+            print("  JWT validation: enabled (no introspection endpoint provided)")
+        } else {
+            print("  Token introspection: enabled")
+        }
+        if let audience = jsonConfig.audience {
+            print("  Expected audience: \(audience)")
+        }
+        if let clientID = jsonConfig.clientID {
+            print("  Expected client ID: \(clientID)")
+        }
+    }
+
+    private static func makeTokenAuthorizationHandler(
+        requiredToken: String
+    ) -> HTTPSSETransport.AuthorizationHandler {
+        return { token in
+            guard let token else {
+                return .unauthorized("Missing bearer token")
+            }
+            guard token == requiredToken else {
+                return .unauthorized("Invalid bearer token")
+            }
+            return .authorized
+        }
+    }
+
+    private func startTCPTransportIfNeeded(server: any MCPServer) async throws -> TCPBonjourTransport? {
+        guard tcp else { return nil }
+        let transport = TCPBonjourTransport(server: server)
+        try await transport.start()
+        print("MCP Server \(server.serverName) started with TCP+Bonjour transport")
+        return transport
+    }
+
+    private func setupSignalHandling(
+        httpTransport: HTTPSSETransport,
+        tcpTransport: TCPBonjourTransport?
+    ) async {
+        if let tcpTransport {
+            signalHandler = SignalHandler(transports: [httpTransport, tcpTransport])
+        } else {
+            signalHandler = SignalHandler(transport: httpTransport)
+        }
+        await signalHandler?.setup()
     }
 }
