+++ Tests/SwiftMCPTests/ComplexTypesTests.swift
@@ -225,483 +225,4 @@ class ComplexTypesServer {
     }
 }
 
-// MARK: - Tests
-
-// MARK: - Basic Type Tests
-
-@Test("Tests processing of integer arrays")
-func testIntArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processIntArray",
-            "arguments": [
-                "numbers": [1, 2, 3, 4, 5]
-            ]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[2,4,6,8,10]")
-}
-
-@Test("Tests processing of optional integer arrays")
-func testOptionalIntArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processOptionalIntArray",
-            "arguments": [:]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[]")
-}
-
-@Test("Tests processing of string arrays")
-func testStringArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processStringArray",
-            "arguments": [
-                "strings": ["hello", "world", "test"]
-            ]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[\"HELLO\",\"WORLD\",\"TEST\"]")
-}
-
-@Test("Tests processing of double arrays")
-func testDoubleArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processDoubleArray",
-            "arguments": [
-                "numbers": [1.1, 2.2, 3.3]
-            ]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[2.2,4.4,6.6]")
-}
-
-@Test("Tests processing of optional double arrays")
-func testOptionalDoubleArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processOptionalDoubleArray",
-            "arguments": [
-                "numbers": [1.1, 2.2, 3.3]
-            ]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[2.2,4.4,6.6]")
-}
-
-@Test("Tests processing of boolean arrays")
-func testBooleanArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processBooleanArray",
-            "arguments": [
-                "values": [true, false, true]
-            ]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[false,true,false]")
-}
-
-@Test("Tests processing of optional boolean arrays")
-func testOptionalBooleanArrayProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    let request = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processOptionalBooleanArray",
-            "arguments": [
-                "values": [true, false, true]
-            ]
-        ]
-    )
-
-    let message = await client.send(request)
-    guard case .response(let response) = message else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(response.id == .int(1))
-    let result = try #require(response.result)
-    let isError = try #require(result["isError"]?.value as? Bool)
-    #expect(isError == false)
-    let content = try #require(result["content"]?.value as? [[String: String]])
-    let firstContent = try #require(content.first)
-    let type = try #require(firstContent["type"])
-    let text = try #require(firstContent["text"])
-    #expect(type == "text")
-    #expect(text == "[false,true,false]")
-}
-
-// MARK: - Complex Type Tests
-
-@Test("Tests creating and processing contact info")
-func testContactInfoProcessing() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    // First create a contact
-    let createRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "createContact",
-            "arguments": [
-                "name": "John Doe",
-                "email": "john@example.com",
-                "phone": "+1234567890",
-                "age": 30,
-                "isActive": true
-            ]
-        ]
-    )
-
-    let createMessage = await client.send(createRequest)
-    guard case .response(let createResponse) = createMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(createResponse.id == .int(1))
-    let createResult = try #require(createResponse.result)
-    let createIsError = try #require(createResult["isError"]?.value as? Bool)
-    #expect(createIsError == false)
-    let structured = try #require(createResult["structuredContent"]?.value as? [String: Any])
-    #expect(structured["name"] as? String == "John Doe")
-    #expect(structured["email"] as? String == "john@example.com")
-    #expect(structured["phone"] as? String == "+1234567890")
-    #expect(structured["age"] as? Int == 30)
-    #expect(structured["isActive"] as? Bool == true)
-    let createContent = try #require(createResult["content"]?.value as? [[String: String]])
-    let createFirstContent = try #require(createContent.first)
-    let createText = try #require(createFirstContent["text"])
-    let sortedStructuredData = try JSONSerialization.data(withJSONObject: structured, options: [.sortedKeys])
-    let sortedStructuredText = try #require(String(data: sortedStructuredData, encoding: .utf8))
-    #expect(sortedStructuredText == createText)
-
-    // Now process the contact array
-    let processRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processContactArray",
-            "arguments": .object([
-                "contacts": .array([.string(createText)])
-            ])
-        ]
-    )
-
-    let processMessage = await client.send(processRequest)
-    guard case .response(let processResponse) = processMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(processResponse.id == .int(1))
-    let processResult = try #require(processResponse.result)
-    let processIsError = try #require(processResult["isError"]?.value as? Bool)
-    #expect(processIsError == false)
-    let processContent = try #require(processResult["content"]?.value as? [[String: String]])
-    let processFirstContent = try #require(processContent.first)
-    let processText = try #require(processFirstContent["text"])
-
-    // Verify the processed contact
-    let json = (processText as String).data(using: String.Encoding.utf8)!
-    let processedWrapper = try JSONDecoder().decode(ContactInfoArrayOutput.self, from: json)
-    let processedContact = try #require(processedWrapper.items.first)
-    #expect(processedContact.name == "JOHN DOE")
-    #expect(processedContact.email == "john@example.com")
-    #expect(processedContact.phone == "+1234567890")
-    #expect(processedContact.age == 60)
-    #expect(processedContact.isActive == false)
-}
-
-@Test("Tests creating a complete profile")
-func testProfileCreation() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    // First create a contact
-    let contactRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "createContact",
-            "arguments": [
-                "name": "Jane Doe",
-                "email": "jane@example.com",
-                "phone": "+1987654321",
-                "age": 25,
-                "isActive": true
-            ]
-        ]
-    )
-
-    let contactMessage = await client.send(contactRequest)
-    guard case .response(let contactResponse) = contactMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    let contactResult = try #require(contactResponse.result)
-    let contactContent = try #require(contactResult["content"]?.value as? [[String: String]])
-    let contactText = try #require(contactContent.first?["text"])
-
-    // Create an address
-    let addressRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "createAddress",
-            "arguments": [
-                "street": "123 Main St",
-                "city": "New York",
-                "postalCode": "10001",
-                "country": "USA"
-            ]
-        ]
-    )
-
-    let addressMessage = await client.send(addressRequest)
-    guard case .response(let addressResponse) = addressMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    let addressResult = try #require(addressResponse.result)
-    let addressContent = try #require(addressResult["content"]?.value as? [[String: String]])
-    let addressText = try #require(addressContent.first?["text"])
-
-    // Create the profile
-    let profileRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "createProfile",
-            "arguments": .object([
-                "contact": .string(contactText),
-                "address": .string(addressText),
-                "interests": .array(["reading", "gaming", "coding"]),
-                "scores": .array([95, 88, 92]),
-                "ratings": .array([4.5, 4.8, 4.2]),
-                "activeStatuses": .array([true, false, true])
-            ])
-        ]
-    )
-
-    let profileMessage = await client.send(profileRequest)
-    guard case .response(let profileResponse) = profileMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    #expect(profileResponse.id == .int(1))
-    let profileResult = try #require(profileResponse.result)
-    let profileIsError = try #require(profileResult["isError"]?.value as? Bool)
-    #expect(profileIsError == false)
-    let profileContent = try #require(profileResult["content"]?.value as? [[String: String]])
-    let profileFirstContent = try #require(profileContent.first)
-    let profileText: String = try #require(profileFirstContent["text"])
-    let profile = try JSONDecoder().decode(Profile.self, from: profileText.data(using: String.Encoding.utf8)!)
-    #expect(profile.contact.name == "Jane Doe")
-    #expect(profile.contact.email == "jane@example.com")
-    #expect(profile.contact.phone == "+1987654321")
-    #expect(profile.contact.age == 25)
-    #expect(profile.contact.isActive == true)
-    #expect(profile.address.street == "123 Main St")
-    #expect(profile.address.city == "New York")
-    #expect(profile.address.postalCode == "10001")
-    #expect(profile.address.country == "USA")
-    #expect(profile.interests == ["reading", "gaming", "coding"])
-    #expect(profile.scores == [95, 88, 92])
-    #expect(profile.ratings == [4.5, 4.8, 4.2])
-    #expect(profile.activeStatuses == [true, false, true])
-}
-
-// MARK: - Optional Array Tests
-
-@Test("Tests processing of optional arrays with nil value")
-func testOptionalArraysWithNil() async throws {
-    let server = ComplexTypesServer()
-    let client = MockClient(server: server)
-
-    // Test optional int array with nil
-    let intRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processOptionalIntArray",
-            "arguments": [:]
-        ]
-    )
-
-    let intMessage = await client.send(intRequest)
-    guard case .response(let intResponse) = intMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    let intResult = try #require(intResponse.result)
-    let intContent = try #require(intResult["content"]?.value as? [[String: String]])
-    let intText = try #require(intContent.first?["text"])
-    #expect(intText == "[]")
-
-    // Test optional string array with nil
-    let stringRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processOptionalStringArray",
-            "arguments": [:]
-        ]
-    )
-
-    let stringMessage = await client.send(stringRequest)
-    guard case .response(let stringResponse) = stringMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    let stringResult = try #require(stringResponse.result)
-    let stringContent = try #require(stringResult["content"]?.value as? [[String: String]])
-    let stringText = try #require(stringContent.first?["text"])
-    #expect(stringText == "[]")
-
-    // Test optional contact array with nil
-    let contactRequest = JSONRPCMessage.request(
-        id: 1,
-        method: "tools/call",
-        params: [
-            "name": "processOptionalContactArray",
-            "arguments": [:]
-        ]
-    )
-
-    let contactMessage = await client.send(contactRequest)
-    guard case .response(let contactResponse) = contactMessage else {
-        #expect(Bool(false), "Expected response case")
-        return
-    }
-    let contactResult = try #require(contactResponse.result)
-    let contactContent = try #require(contactResult["content"]?.value as? [[String: String]])
-    let contactText = try #require(contactContent.first?["text"])
-    #expect(contactText == "{\"items\":[]}")
-}
+// Tests live in ComplexTypesArrayTests.swift and ComplexTypesObjectTests.swift
