+++ Tests/SwiftMCPTests/MCPToolTests.swift
@@ -106,200 +106,115 @@ class URLParameterHandling {
     }
 }
 
-// Test class with SchemaRepresentable types
-@MCPServer
-class SchemaRepresentableTests {
-    /// A person's contact information
-    @Schema
-    struct ContactInfo {
-        /// The person's full name
-        let name: String
-
-        /// The person's email address
-        let email: String
-
-        /// The person's phone number (optional)
-        let phone: String?
-
-        /// The person's age
-        var age: Int = 0
-
-        /// The person's address
-        let address: Address
-    }
-
-    /// A person's address
-    @Schema
-    struct Address: Codable {
-        let street: String
-        let city: String
-        let zip: String
-    }
-
-    /**
-     Get reminders from the reminders app with flexible filtering options.
-     
-     - Parameters:
-        - contact: A test contact
-     */
-    @MCPTool
-    func fetchReminders(
-        contact: Address
-    ) -> String {
-        return "\(contact)"
-    }
-}
-
-// Test class with array of enums
-@MCPServer
-class EnumArrayTest {
-    /// Function that takes an array of weekdays
-    /// - Parameter days: Array of weekdays
-    @MCPTool
-    func processWeekdays(days: [Weekday]) {
-        // Implementation not important for the test
-    }
-
-    /// Function that takes an optional array of weekdays
-    /// - Parameter days: Optional array of weekdays
-    @MCPTool
-    func processOptionalWeekdays(days: [Weekday]? = nil) {
-        // Implementation not important for the test
-    }
-}
-
-// Test class with array of SchemaRepresentable types
-@MCPServer
-class SchemaRepresentableArrayTest {
-    /// Function that takes an array of addresses
-    /// - Parameter addresses: Array of addresses
-    @MCPTool
-    func processAddresses(addresses: [SchemaRepresentableTests.Address]) {
-        // Implementation not important for the test
-    }
-}
-
-// Add Weekday enum before the tests
-enum Weekday: String, CaseIterable {
-    case monday
-    case tuesday
-    case wednesday
-    case thursday
-    case friday
-    case saturday
-    case sunday
-}
+// Test fixture classes live in MCPToolTestFixtures.swift
 
 // MARK: - Tests
 
+private func assertNoParametersTool(_ tools: [MCPTool]) {
+    guard let noParamsTool = tools.first(where: { $0.name == "noParameters" }) else {
+        #expect(Bool(false), "Could not find noParameters function")
+        return
+    }
+    #expect(noParamsTool.description == "Simple function with no parameters")
+    if case .object(let object, _) = noParamsTool.inputSchema {
+        #expect(object.properties.isEmpty)
+    } else {
+        #expect(Bool(false), "Expected object schema")
+    }
+}
+
+private func assertBasicTypesParameters(_ object: JSONSchema.Object) {
+    #expect(object.properties.count == 3)
+    if case .number(title: _, description: let description, minimum: _, maximum: _, defaultValue: _) =
+        object.properties["a"] {
+        #expect(description == "An integer parameter")
+    } else {
+        #expect(Bool(false), "Expected number schema for parameter 'a'")
+    }
+    if case .string(
+        title: _,
+        description: let description,
+        format: _,
+        minLength: _,
+        maxLength: _,
+        defaultValue: _
+    ) = object.properties["b"] {
+        #expect(description == "A string parameter")
+    } else {
+        #expect(Bool(false), "Expected string schema for parameter 'b'")
+    }
+    if case .boolean(title: _, description: let description, defaultValue: _) = object.properties["c"] {
+        #expect(description == "A boolean parameter")
+    } else {
+        #expect(Bool(false), "Expected boolean schema for parameter 'c'")
+    }
+}
+
+private func assertBasicTypesTool(_ tools: [MCPTool]) {
+    guard let basicTypesTool = tools.first(where: { $0.name == "basicTypes" }) else {
+        #expect(Bool(false), "Could not find basicTypes function")
+        return
+    }
+    #expect(basicTypesTool.description == "Function with basic parameter types")
+    if case .object(let object, _) = basicTypesTool.inputSchema {
+        assertBasicTypesParameters(object)
+    } else {
+        #expect(Bool(false), "Expected object schema")
+    }
+}
+
+private func assertExplicitDescriptionTool(_ tools: [MCPTool]) {
+    guard let explicitDescTool = tools.first(where: { $0.name == "explicitDescription" }) else {
+        #expect(Bool(false), "Could not find explicitDescription function")
+        return
+    }
+    #expect(explicitDescTool.description == "This description overrides the documentation comment")
+    if case .object(let object, _) = explicitDescTool.inputSchema {
+        if case .number(title: _, description: let description, minimum: _, maximum: _, defaultValue: _) =
+            object.properties["value"] {
+            #expect(description == "A value with description")
+        } else {
+            #expect(Bool(false), "Expected number schema for parameter 'value'")
+        }
+    } else {
+        #expect(Bool(false), "Expected object schema")
+    }
+}
+
+private func assertOptionalParameterTool(_ tools: [MCPTool]) {
+    guard let optionalParamTool = tools.first(where: { $0.name == "optionalParameter" }) else {
+        #expect(Bool(false), "Could not find optionalParameter function")
+        return
+    }
+    #expect(optionalParamTool.description == "Function with optional parameters")
+    if case .object(let object, _) = optionalParamTool.inputSchema {
+        if case .string(
+            title: _, description: let description, format: _, minLength: _, maxLength: _, defaultValue: _
+        ) = object.properties["required"] {
+            #expect(description == "A required parameter")
+        } else {
+            #expect(Bool(false), "Expected string schema for parameter 'required'")
+        }
+        if case .number(title: _, description: let description, minimum: _, maximum: _, defaultValue: _) =
+            object.properties["optional"] {
+            #expect(description == "An optional parameter")
+        } else {
+            #expect(Bool(false), "Expected string schema for parameter 'optional'")
+        }
+    } else {
+        #expect(Bool(false), "Expected object schema")
+    }
+}
+
 @Test
 func testBasicFunctionality() {
-    // Create an instance of the test class
     let instance = TripleSlashDocumentation()
-
-    // Get the tools array
     let tools = instance.mcpToolMetadata.convertedToTools()
 
-    // Test that the tools array contains the expected functions
-    if let noParamsTool = tools.first(where: { $0.name == "noParameters" }) {
-        #expect(noParamsTool.description == "Simple function with no parameters")
-
-        // Extract properties from the object schema
-        if case .object(let object, _) = noParamsTool.inputSchema {
-			#expect(object.properties.isEmpty)
-        } else {
-            #expect(Bool(false), "Expected object schema")
-        }
-    } else {
-        #expect(Bool(false), "Could not find noParameters function")
-    }
-
-    // Test basic parameter types
-    if let basicTypesTool = tools.first(where: { $0.name == "basicTypes" }) {
-        #expect(basicTypesTool.description == "Function with basic parameter types")
-
-        // Extract properties from the object schema
-        if case .object(let object, _) = basicTypesTool.inputSchema {
-			#expect(object.properties.count == 3)
-
-            // Check parameter descriptions
-			if case .number(title: _, description: let description, minimum: _, maximum: _, defaultValue: _) =
-                object.properties["a"] {
-                #expect(description == "An integer parameter")
-            } else {
-                #expect(Bool(false), "Expected number schema for parameter 'a'")
-            }
-
-			if case .string(
-                title: _,
-                description: let description,
-                format: _,
-                minLength: _,
-                maxLength: _,
-                defaultValue: _
-            ) = object.properties["b"] {
-                #expect(description == "A string parameter")
-            } else {
-                #expect(Bool(false), "Expected string schema for parameter 'b'")
-            }
-
-			if case .boolean(title: _, description: let description, defaultValue: _) = object.properties["c"] {
-                #expect(description == "A boolean parameter")
-            } else {
-                #expect(Bool(false), "Expected boolean schema for parameter 'c'")
-            }
-        } else {
-            #expect(Bool(false), "Expected object schema")
-        }
-    } else {
-        #expect(Bool(false), "Could not find basicTypes function")
-    }
-
-    // Test explicit description override
-    if let explicitDescTool = tools.first(where: { $0.name == "explicitDescription" }) {
-        #expect(explicitDescTool.description == "This description overrides the documentation comment")
-
-        // Extract properties from the object schema
-        if case .object(let object, _) = explicitDescTool.inputSchema {
-			if case .number(title: _, description: let description, minimum: _, maximum: _, defaultValue: _) =
-                object.properties["value"] {
-                #expect(description == "A value with description")
-            } else {
-                #expect(Bool(false), "Expected number schema for parameter 'value'")
-            }
-        } else {
-            #expect(Bool(false), "Expected object schema")
-        }
-    } else {
-        #expect(Bool(false), "Could not find explicitDescription function")
-    }
-
-    // Test optional parameters
-    if let optionalParamTool = tools.first(where: { $0.name == "optionalParameter" }) {
-        #expect(optionalParamTool.description == "Function with optional parameters")
-
-        // Extract properties from the object schema
-        if case .object(let object, _) = optionalParamTool.inputSchema {
-			if case .string(title: _, description: let description, format: _, minLength: _, maxLength: _, defaultValue: _) =
-                object.properties["required"] {
-                #expect(description == "A required parameter")
-            } else {
-                #expect(Bool(false), "Expected string schema for parameter 'required'")
-            }
-
-            // Optional parameters are represented as strings in the schema
-			if case .number(title: _, description: let description, minimum: _, maximum: _, defaultValue: _) =
-                object.properties["optional"] {
-                #expect(description == "An optional parameter")
-            } else {
-                #expect(Bool(false), "Expected string schema for parameter 'optional'")
-            }
-        } else {
-            #expect(Bool(false), "Expected object schema")
-        }
-    } else {
-        #expect(Bool(false), "Could not find optionalParameter function")
-    }
+    assertNoParametersTool(tools)
+    assertBasicTypesTool(tools)
+    assertExplicitDescriptionTool(tools)
+    assertOptionalParameterTool(tools)
 }
 
 @Test
@@ -413,283 +328,4 @@ func testURLParameters() async throws {
     }
 }
 
-@Test
-func testSchemaRepresentableParameter() async throws {
-    let instance = SchemaRepresentableTests()
-    let tools = instance.mcpToolMetadata.convertedToTools()
-
-    // Create test data
-    let address = SchemaRepresentableTests.Address(street: "123 Main St", city: "New York", zip: "10001")
-
-    // Create parameters dictionary
-    let params: JSONDictionary = [
-        "contact": try JSONValue(encoding: address)
-    ]
-
-    // Call the function
-    let result = try await instance.callTool("fetchReminders", arguments: params)
-
-    // Verify the result
-    #expect(result as? String == "Address(street: \"123 Main St\", city: \"New York\", zip: \"10001\")")
-
-    // Verify the schema
-    if let tool = tools.first(where: { $0.name == "fetchReminders" }) {
-        // Verify the schema matches the expected JSON schema
-        if case .object(let object, _) = tool.inputSchema {
-			#expect(object.properties.count == 1)
-			#expect(object.required == ["contact"])
-
-            // Verify the contact property
-			if case .object(let object, _) = object.properties["contact"] {
-				#expect(object.properties.count == 3)
-				#expect(object.required == ["street", "city", "zip"])
-
-                // Verify property types
-				if case .string = object.properties["street"] {
-                    // street property is correct
-                } else {
-                    #expect(Bool(false), "Expected string schema for street property")
-                }
-
-				if case .string = object.properties["city"] {
-                    // city property is correct
-                } else {
-                    #expect(Bool(false), "Expected string schema for city property")
-                }
-
-				if case .string = object.properties["zip"] {
-                    // zip property is correct
-                } else {
-                    #expect(Bool(false), "Expected string schema for zip property")
-                }
-            } else {
-                #expect(Bool(false), "Expected object schema for contact property")
-            }
-        } else {
-            #expect(Bool(false), "Expected object schema")
-        }
-    } else {
-        #expect(Bool(false), "Could not find fetchReminders function")
-    }
-}
-
-@Test
-func testEnumArraySchema() throws {
-    let server = EnumArrayTest()
-    let tools = server.mcpToolMetadata.convertedToTools()
-
-    // Find the processWeekdays tool
-    guard let tool = tools.first(where: { $0.name == "processWeekdays" }) else {
-        #expect(Bool(false), "Could not find processWeekdays tool")
-        return
-    }
-
-    // Get the schema for the days parameter
-    if case .object(let object, _) = tool.inputSchema {
-		guard let daysSchema = object.properties["days"] else {
-            #expect(Bool(false), "Could not find days parameter in schema")
-            return
-        }
-
-        // Verify it's an array
-        if case .array(let itemsSchema, title: _, description: _, defaultValue: _) = daysSchema {
-            // Verify the items are strings with enum values
-            if case .enum(let enumValues, title: _, description: _, enumNames: _, defaultValue: _) = itemsSchema {
-
-                // Verify the enum values are the Weekday cases
-                let expectedValues = ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
-                #expect(enumValues.sorted() == expectedValues.sorted())
-            } else {
-                #expect(Bool(false), "Array items should be strings")
-            }
-        } else {
-            #expect(Bool(false), "Expected array schema")
-        }
-    } else {
-        #expect(Bool(false), "Expected object schema")
-    }
-}
-
-@Test
-func testOptionalEnumArraySchema() throws {
-    let server = EnumArrayTest()
-    let tools = server.mcpToolMetadata.convertedToTools()
-
-    // Find the processOptionalWeekdays tool
-    guard let tool = tools.first(where: { $0.name == "processOptionalWeekdays" }) else {
-        #expect(Bool(false), "Could not find processOptionalWeekdays tool")
-        return
-    }
-
-    // Get the schema for the days parameter
-    if case .object(let object, _) = tool.inputSchema {
-		guard let daysSchema = object.properties["days"] else {
-            #expect(Bool(false), "Could not find days parameter in schema")
-            return
-        }
-
-        // Verify it's not in the required array
-		#expect(!object.required.contains("days"), "Optional parameter should not be required")
-
-        // Verify it's an array
-        if case .array(let itemsSchema, title: _, description: _, defaultValue: _) = daysSchema {
-            // Verify the items are strings with enum values
-            if case .enum(let enumValues, title: _, description: _, enumNames: _, defaultValue: _) = itemsSchema {
-
-                // Verify the enum values are the Weekday cases
-                let expectedValues = ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
-                #expect(enumValues.sorted() == expectedValues.sorted())
-            } else {
-                #expect(Bool(false), "Array items should be strings")
-            }
-        } else {
-            #expect(Bool(false), "Expected array schema")
-        }
-    } else {
-        #expect(Bool(false), "Expected object schema")
-    }
-}
-
-@Test
-func testSchemaRepresentableArraySchema() throws {
-    let server = SchemaRepresentableArrayTest()
-    let tools = server.mcpToolMetadata.convertedToTools()
-
-    // Find the processAddresses tool
-    guard let tool = tools.first(where: { $0.name == "processAddresses" }) else {
-        #expect(Bool(false), "Could not find processAddresses tool")
-        return
-    }
-
-    // Get the schema for the addresses parameter
-    if case .object(let object, _) = tool.inputSchema {
-		guard let addressesSchema = object.properties["addresses"] else {
-            #expect(Bool(false), "Could not find addresses parameter in schema")
-            return
-        }
-
-        // Verify it's required
-		#expect(object.required.contains("addresses"), "Required parameter should be in required array")
-
-        // Verify it's an array
-        if case .array(let itemsSchema, title: _, description: _, defaultValue: _) = addressesSchema {
-            // Verify the items are objects with street and city properties
-            if case .object(let object, _) = itemsSchema {
-				#expect(object.properties.count == 3)
-				#expect(object.required == ["street", "city", "zip"])
-
-                // Verify property types
-				if case .string = object.properties["street"] {
-                    // street property is correct
-                } else {
-                    #expect(Bool(false), "Expected string schema for street property")
-                }
-
-				if case .string = object.properties["city"] {
-                    // city property is correct
-                } else {
-                    #expect(Bool(false), "Expected string schema for city property")
-                }
-
-				if case .string = object.properties["zip"] {
-                    // zip property is correct
-                } else {
-                    #expect(Bool(false), "Expected string schema for zip property")
-                }
-            } else {
-                #expect(Bool(false), "Expected object schema for array items")
-            }
-        } else {
-            #expect(Bool(false), "Expected array schema")
-        }
-    } else {
-        #expect(Bool(false), "Expected object schema")
-    }
-}
-
-@Test
-func testAddressDecoding() throws {
-    // Create a dictionary representing an address
-    let addressDict: [String: Any] = [
-        "street": "123 Main St",
-        "city": "New York",
-        "zip": "10001"
-    ]
-
-    // Encode the dictionary to JSON data
-    let jsonData = try JSONSerialization.data(withJSONObject: addressDict)
-
-    // Decode the JSON data into an Address struct
-    let address = try JSONDecoder().decode(SchemaRepresentableTests.Address.self, from: jsonData)
-
-    // Verify the decoded values
-    #expect(address.street == "123 Main St")
-    #expect(address.city == "New York")
-    #expect(address.zip == "10001")
-}
-
-@Test
-func testAddressArrayDecoding() throws {
-    // Create an array of dictionaries representing addresses
-    let addressDicts: [[String: Any]] = [
-        [
-            "street": "123 Main St",
-            "city": "New York",
-            "zip": "10001"
-        ],
-        [
-            "street": "456 Oak Ave",
-            "city": "San Francisco",
-            "zip": "94102"
-        ],
-        [
-            "street": "789 Pine Rd",
-            "city": "Chicago",
-            "zip": "60601"
-        ]
-    ]
-
-    // Encode the array to JSON data
-    let jsonData = try JSONSerialization.data(withJSONObject: addressDicts)
-
-    // Decode the JSON data into an array of Address structs
-    let addresses = try JSONDecoder().decode([SchemaRepresentableTests.Address].self, from: jsonData)
-
-    // Verify the number of addresses
-    #expect(addresses.count == 3)
-
-    // Verify each address
-    #expect(addresses[0].street == "123 Main St")
-    #expect(addresses[0].city == "New York")
-    #expect(addresses[0].zip == "10001")
-
-    #expect(addresses[1].street == "456 Oak Ave")
-    #expect(addresses[1].city == "San Francisco")
-    #expect(addresses[1].zip == "94102")
-
-    #expect(addresses[2].street == "789 Pine Rd")
-    #expect(addresses[2].city == "Chicago")
-    #expect(addresses[2].zip == "60601")
-}
-
-@Test
-func testExtractAddressArray() throws {
-    let addresses = [
-        SchemaRepresentableTests.Address(street: "123 Main St", city: "New York", zip: "10001"),
-        SchemaRepresentableTests.Address(street: "456 Oak Ave", city: "San Francisco", zip: "94102")
-    ]
-
-    let params: JSONDictionary = [
-        "addresses": try JSONValue(encoding: addresses)
-    ]
-
-    let extractedAddresses: [SchemaRepresentableTests.Address] = try params.extractParameter(named: "addresses")
-
-    #expect(extractedAddresses.count == 2)
-    #expect(extractedAddresses[0].street == "123 Main St")
-    #expect(extractedAddresses[0].city == "New York")
-    #expect(extractedAddresses[0].zip == "10001")
-    #expect(extractedAddresses[1].street == "456 Oak Ave")
-    #expect(extractedAddresses[1].city == "San Francisco")
-    #expect(extractedAddresses[1].zip == "94102")
-}
+// Schema tests live in MCPToolSchemaTests.swift
